requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
chromadb==0.4.24
python-dotenv==1.0.1
langchain==0.1.13
ollama==0.1.7
PyPDF2==3.0.1
python-docx==1.1.0